# apps/api/src/purchase.ts
import { Body, Controller, Get, Injectable, Module, Param, Post, Req, Res, UseGuards } from '@nestjs/common';
import type { FastifyReply, FastifyRequest } from 'fastify';
import { newId, rows, sql, withTenant, type Transaction } from '@ecommerce/database';
import { requestRefund, consumerReply, ticket, allocate, attemptData, cancel, checkoutExisting, correctAddress, createCheckout, newAttempt, decrypt, encrypt, event, hash, order, protocol, PurchaseError, receiveWebhook, requireSimulation, signature, simulationEnabled, view } from '@ecommerce/purchase';
import { AccessModule, SessionGuard, type AuthRequest } from './access.js';
import { CartModule, CartService } from './cart.js';
import { id, object, text } from './catalogue.js';
import { Infrastructure } from './infrastructure.js';
import { StoresModule, StoresService } from './stores.js';
import { randomToken } from './security.js';
import { MfaModule, requireMfa } from './mfa.js';
// Buyers see their receipt and blocking state, never internal notes, history actors or provider identifiers.
async function buyerView(tx:Transaction,orderId:string){const {history,incidents,attempts,transactions,address_corrections,outbox_failures,principal_id,cart_id,cart_version,version,...o}=await view(tx,orderId) as any;return {...o,attempts:attempts.map((a:any)=>({id:a.id,method:a.method,status:a.status,environment:a.environment})),transactions:transactions.map((t:any)=>({id:t.id,received_cents:t.received_cents,classification:t.classification,refunded_cents:t.refunded_cents})),incidents:incidents.map((i:any)=>({id:i.id,code:i.code,status:i.status}))};}
@Injectable()
class PurchaseService {
 constructor(private cart:CartService,private infra:Infrastructure,private stores:StoresService){}
 async checkout(slug:string,req:FastifyRequest,reply:FastifyReply,input:unknown){try{return await this.confirm(slug,req,reply,input);}catch(error){await this.failure(slug,error);throw error;}}
 // Pilot observation: reason code and status only, recorded outside the failed transaction; recording never masks the original error.
 private async failure(slug:string,error:unknown){try{const status=(error as {status?:number}).status??(error as {getStatus?:()=>number}).getStatus?.()??500,message=String((error as Error).message||''),reason=/desatualizado|Cotação/.test(message)?'PRICE_OR_SHIPPING_CHANGED':/Saldo|saldo/.test(message)?'STOCK_UNAVAILABLE':/suspensas|Vitrine indisponível/.test(message)?'SALES_CLOSED':/conta|Conta|Fornecedor/.test(message)?'PAYMENT_ACCOUNT':/Chave reutilizada/.test(message)?'IDEMPOTENCY_CONFLICT':status===400?'INVALID_INPUT':status===503?'SERVICE_UNAVAILABLE':'OTHER';
  const [route]=(await this.infra.shop.pool.query('select tenant_id from shop.resolve_store($1,false)',[slug])).rows;if(route)await withTenant(this.infra.shop,route.tenant_id,null,tx=>tx.execute(sql`insert into shop.checkout_failures(id,tenant_id,reason,status) values(${newId()},${route.tenant_id},${reason},${status})`));}catch{console.error(JSON.stringify({event:'checkout_failure_unrecorded'}));}}
 private async confirm(slug:string,req:FastifyRequest,reply:FastifyReply,input:unknown){requireSimulation();const b=object(input),key=text(b.key,100),quoteId=id(b.quote_id),address=this.cart.address(b.address),buyerInput=object(b.buyer),buyer={name:text(buyerInput.name,100),email:text(buyerInput.email,200).toLowerCase()},method=text(b.method,10),total=text(b.total_cents,20);if(!/^[1-9]\d*$/.test(total)||BigInt(total)>9223372036854775807n||!['PIX','CARD'].includes(method)||!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(buyer.email))throw new PurchaseError(400,'Revise comprador, método e total.');const fingerprint=hash(JSON.stringify({quoteId,address,buyer,method,total}));
 // Fail closed for new checkout when queue recovery infrastructure is unavailable.
 try{await this.infra.queue.getJobCounts('wait');}catch{throw new PurchaseError(503,'Processamento indisponível; tente depois.');}
 return this.cart.run(slug,req,reply,async(tx,cart,route)=>{const existing=await checkoutExisting(tx,cart,key,fingerprint);if(existing)return buyerView(tx,existing);
 await tx.execute(sql`select v.id from shop.product_variants v join shop.cart_items c on (c.tenant_id,c.variant_id)=(v.tenant_id,v.id) where c.cart_id=${cart.id} order by v.id for share of v`);
 const data=await this.cart.items(tx,cart.id);const [quote]=await rows<{fingerprint:string;rule_id:string;version:string;price_cents:string;name:string;kind:string;days:number}>(tx,sql`select q.fingerprint,q.rule_id,r.version::text,q.price_cents::text,case when q.provider_service is not null then r.name||' — '||q.provider_service else r.name end as name,r.kind,coalesce(q.days,r.days) as days from shop.shipping_quotes q join shop.shipping_rules r on (r.tenant_id,r.id)=(q.tenant_id,q.rule_id) where q.id=${quoteId} and q.cart_id=${cart.id} and q.expires_at>now() and r.active for share of r`);
 if(!quote||!data.items.length||!data.valid||quote.fingerprint!==hash(JSON.stringify({address,items:data.items,rule:quote.rule_id,version:quote.version}))||(BigInt(data.subtotal_cents)+BigInt(quote.price_cents)).toString()!==total)throw new PurchaseError(409,'Preço/frete desatualizado ou indisponível. Revise e confirme novamente.');
 const [profile]=await rows<{profile:unknown}>(tx,sql`select profile from shop.merchant_profiles`);if(!profile)throw new PurchaseError(409,'Fornecedor não configurado.');const orderId=await createCheckout(tx,route.tenant_id,cart,{key,fingerprint,items:data.items,subtotal:data.subtotal_cents,shipping:quote.price_cents,address,buyer,supplier:profile.profile,shippingSnapshot:{name:quote.name,kind:quote.kind,days:quote.days,rule_id:quote.rule_id,version:quote.version},method:method as 'PIX'|'CARD'});return buyerView(tx,orderId);});}
 async buyer<T>(slug:string,req:FastifyRequest,orderId:string,action:(tx:Transaction,tenant:string)=>Promise<T>){const [route]=(await this.infra.shop.pool.query('select * from shop.resolve_store($1,false)',[slug])).rows;if(!route)throw new PurchaseError(404,'Pedido não autorizado.');if(req.method!=='GET'&&req.headers.origin!==process.env.PUBLIC_ORIGIN&&req.headers.origin!==route.canonical)throw new PurchaseError(403,'Origem inválida.');const cookie=req.cookies[`cart_${route.tenant_id.replaceAll('-','')}`],header=req.headers['x-order-token'],link=typeof header==='string'&&/^[A-Za-z0-9_-]{40,60}$/.test(header)?header:'';if(!cookie&&!link)throw new PurchaseError(404,'Pedido não autorizado.');return withTenant(this.infra.shop,route.tenant_id,null,async tx=>{
 // Either the cart cookie of the purchase or an unexpired, unrevoked e-mail link secret (sent in the URL fragment, never in the path).
 const [authorized]=await rows(tx,sql`select o.id from shop.orders o join shop.carts c on (c.tenant_id,c.id)=(o.tenant_id,o.cart_id) where o.id=${orderId} and (c.token_hash=${hash(cookie||'-')} or exists(select 1 from shop.order_access_tokens t where t.order_id=o.id and t.token_hash=${hash(link||'-')} and t.revoked_at is null and t.expires_at>now()))`);if(!authorized)throw new PurchaseError(404,'Pedido não autorizado.');return action(tx,route.tenant_id);});}
 read(slug:string,req:FastifyRequest,orderId:string){return this.buyer(slug,req,id(orderId),tx=>buyerView(tx,orderId));}
 request(slug:string,req:FastifyRequest,orderId:string,input:unknown){const b=object(input),kind=text(b.kind,20),message=text(b.message,2000),key=text(b.key,100);if(!['SUPPORT','WITHDRAWAL','CANCELLATION'].includes(kind))throw new PurchaseError(400,'Tipo inválido.');return this.buyer(slug,req,id(orderId),(tx,tenant)=>protocol(tx,tenant,orderId,key,kind,message));}
 retry(slug:string,req:FastifyRequest,orderId:string,input:unknown){requireSimulation();const b=object(input),method=text(b.method,10),key=text(b.key,100);if(!['PIX','CARD'].includes(method))throw new PurchaseError(400,'Método inválido.');return this.buyer(slug,req,id(orderId),async(tx,tenant)=>{await newAttempt(tx,tenant,orderId,method as 'PIX'|'CARD',key);return buyerView(tx,orderId);});}
 requestView(slug:string,req:FastifyRequest,orderId:string,requestId:string){return this.buyer(slug,req,id(orderId),async tx=>{const [r]=await rows(tx,sql`select id from shop.consumer_requests where id=${id(requestId)} and order_id=${orderId}`);if(!r)throw new PurchaseError(404,'Protocolo não encontrado.');return ticket(tx,requestId,false);});}
 requestReply(slug:string,req:FastifyRequest,orderId:string,requestId:string,input:unknown){const body=text(object(input).body,4000);return this.buyer(slug,req,id(orderId),async(tx,tenant)=>{const [r]=await rows(tx,sql`select id from shop.consumer_requests where id=${id(requestId)} and order_id=${orderId}`);if(!r)throw new PurchaseError(404,'Protocolo não encontrado.');return consumerReply(tx,tenant,requestId,body);});}
 async methods(){return {simulation:simulationEnabled(),real_enabled:false,oauth_enabled:false,commission_enabled:false,methods:simulationEnabled()?['PIX_SIMULATED','CARD_SIMULATED']:[],reason:'Homologação externa da Fase 0 pendente; sem movimentação financeira real.'};}
 accounts(tenant:string,req:AuthRequest){return this.stores.run(id(tenant),req.actor,tx=>rows(tx,sql`select id,provider,environment,seller_id,status,created_at from shop.payment_accounts`),true);}
 connect(tenant:string,req:AuthRequest){requireSimulation();return this.stores.run(id(tenant),req.actor,async tx=>{const accountId=newId(),secret=randomToken();const [account]=await rows(tx,sql`insert into shop.payment_accounts(id,tenant_id,provider,environment,seller_id,status,credentials_cipher) values(${accountId},${tenant},'SIMULATED','SIMULATED',${'sim-'+tenant},'CONNECTED',${encrypt(secret,`${tenant}:${accountId}`)}) on conflict(tenant_id,provider,environment) do update set status='CONNECTED' returning id,provider,environment,seller_id,status`);return account;},true);}
 revoke(tenant:string,req:AuthRequest,account:string){return this.stores.run(id(tenant),req.actor,async tx=>{const [a]=await rows(tx,sql`update shop.payment_accounts set status='REVOKED' where id=${id(account)} returning id,status`);if(!a)throw new PurchaseError(404,'Conta não encontrada.');return a;},true);}
 list(tenant:string,req:AuthRequest){return this.stores.run(id(tenant),req.actor,tx=>rows(tx,sql`select id,number::text,order_status,payment_status,fulfillment_status,dispute_status,total_cents::text,created_at from shop.orders order by number desc limit 100`));}
 adminRead(tenant:string,req:AuthRequest,orderId:string){return this.stores.run(id(tenant),req.actor,tx=>view(tx,id(orderId)));}
 cancel(tenant:string,req:AuthRequest,orderId:string,input:unknown){return this.stores.run(id(tenant),req.actor,(tx,member)=>cancel(tx,tenant,id(orderId),member.id,text(object(input).reason,500)),true);}
 note(tenant:string,req:AuthRequest,orderId:string,incidentId:string,input:unknown){return this.stores.run(id(tenant),req.actor,async(tx,member)=>{const o=await order(tx,id(orderId),true);const [row]=await rows(tx,sql`update shop.order_incidents set note=${text(object(input).note,1000)},actor_membership_id=${member.id} where id=${id(incidentId)} and order_id=${o.id} returning id,status,note`);if(!row)throw new PurchaseError(404,'Incidente não encontrado.');await event(tx,tenant,o,'INCIDENT_ANNOTATED','Referência anotada; não confirma devolução.',member.id);return row;},true);}
 address(tenant:string,req:AuthRequest,orderId:string,input:unknown){const b=object(input);return this.stores.run(id(tenant),req.actor,(tx,member)=>correctAddress(tx,tenant,id(orderId),member.id,this.cart.address(b.address),text(b.reason,500)),true);}
 reprocess(tenant:string,req:AuthRequest,orderId:string){return this.stores.run(id(tenant),req.actor,async(tx,member)=>{const o=await order(tx,id(orderId),true);const result=await tx.execute(sql`update shop.purchase_outbox set failures=0,next_run_at=now() where aggregate_id=${o.id} and executed_at is null and failures>0`);await event(tx,tenant,o,'OUTBOX_REPROCESS_REQUESTED','Reprocessamento administrativo de eventos com falha.',member.id);return {requeued:result.rowCount??0};},true);}
 // Platform-initiated refund: Owner with MFA; the gateway call happens after the intent is committed.
 async refund(tenant:string,req:AuthRequest,orderId:string,input:unknown){requireMfa(req.actor);const b=object(input);const member=await this.stores.run(id(tenant),req.actor,async(_tx,m)=>m,true);
  return requestRefund(this.infra.shop,id(tenant),{orderId:id(orderId),transactionId:id(b.transaction_id),amountCents:text(b.amount_cents,20),key:text(b.key,100),reason:text(b.reason,500),member:member.id});}
 reallocate(tenant:string,req:AuthRequest,orderId:string){return this.stores.run(id(tenant),req.actor,async(tx,member)=>{const o=await order(tx,id(orderId),true);const [pending]=await rows(tx,sql`select id from shop.order_incidents where order_id=${o.id} and code='PAID_WITHOUT_STOCK' and status='OPEN'`);if(!pending||o.order_status!=='OPEN'||o.cancel_requested_at||o.payment_status!=='PAID')throw new PurchaseError(409,'Pedido não admite realocação.');const items=await rows<any>(tx,sql`select variant_id,quantity,price_cents::text from shop.order_items where order_id=${o.id}`);if(!await allocate(tx,tenant,o.id,items,true))throw new PurchaseError(409,'Recomponha todo o saldo antes de realocar.');await tx.execute(sql`update shop.order_incidents set status='RESOLVED',resolved_at=now(),actor_membership_id=${member.id} where order_id=${o.id} and code in ('PAID_WITHOUT_STOCK','REFUND_ACTION_REQUIRED') and transaction_id=${o.principal_id}`);await event(tx,tenant,o,'STOCK_REALLOCATED','Estoque recomposto e consumido atomicamente.',member.id);return view(tx,o.id);},true);}
 async simulate(tenant:string,req:AuthRequest,attemptId:string,input:unknown){requireSimulation();const b=object(input),status=text(b.status,20);if(!['PENDING','APPROVED','REJECTED','CANCELLED','EXPIRED','REFUNDED','DISPUTED','DISPUTE_WON','DISPUTE_LOST'].includes(status))throw new PurchaseError(400,'Estado simulado inválido.');
 const info=await this.stores.run(id(tenant),req.actor,async tx=>{const a=await attemptData(tx,id(attemptId));const [f]=await rows<{id:string;amount_cents:string;refunds:any[]}>(tx,sql`select id,amount_cents::text,refunds,dispute from shop.simulated_payments where reference=${a.id} order by created_at limit 1 for update`);if(!f)throw new PurchaseError(409,'Aguarde emissão pela fila.');let refunds=f.refunds;let dispute:any=null;if(status==='REFUNDED'){const amount=text(b.amount_cents,20);if(!/^[1-9]\d*$/.test(amount)||BigInt(amount)+refunds.reduce((s,r)=>s+BigInt(r.amount_cents),0n)>BigInt(f.amount_cents))throw new PurchaseError(400,'Devolução simulada excede recebido.');refunds=[...refunds,{id:newId(),amount_cents:amount,confirmed_at:new Date().toISOString()}];}const current=(f as any).dispute;if(status==='DISPUTED')dispute={id:newId(),status:'OPEN',financial_loss:false};if(status==='DISPUTE_WON'||status==='DISPUTE_LOST'){if(!current)throw new PurchaseError(409,'Sem disputa simulada aberta.');dispute={...current,status:status==='DISPUTE_WON'?'WON':'LOST',financial_loss:status==='DISPUTE_LOST'};}const providerStatus=status.startsWith('DISPUTE_')?(status==='DISPUTE_LOST'?'CHARGED_BACK':'APPROVED'):status;await tx.execute(sql`update shop.simulated_payments set status=${providerStatus},approved_at=case when ${['APPROVED','REFUNDED','DISPUTED','DISPUTE_WON','DISPUTE_LOST'].includes(status)} then coalesce(approved_at,now()) else approved_at end,refunds=${JSON.stringify(refunds)}::jsonb,dispute=coalesce(${dispute?JSON.stringify(dispute):null}::jsonb,dispute),revision=revision+1 where id=${f.id}`);return {a,f};},true);
 const secret=decrypt(info.a.credentials_cipher!,`${tenant}:${info.a.account_id}`),ts=String(Date.now()),eventId=newId();await receiveWebhook(this.infra.shop,info.a.account_id,eventId,info.f.id,eventId,`ts=${ts},v1=${signature(secret,info.f.id,eventId,ts)}`);return {simulation:true,queued:true,message:'Fato do provedor SIMULADO alterado; aguardando consulta autoritativa.'};}
}
@Controller('public/stores/:slug')
class BuyerController {constructor(private service:PurchaseService){}
 @Get('payment-methods') methods(){return this.service.methods();}
 @Post('cart/checkout') checkout(@Param('slug') s:string,@Req() r:FastifyRequest,@Res({passthrough:true}) p:FastifyReply,@Body() b:unknown){return this.service.checkout(s,r,p,b);}
 @Get('orders/:orderId') read(@Param('slug') s:string,@Param('orderId') id:string,@Req() r:FastifyRequest){return this.service.read(s,r,id);}
 @Post('orders/:orderId/attempts') retry(@Param('slug') s:string,@Param('orderId') id:string,@Req() r:FastifyRequest,@Body() b:unknown){return this.service.retry(s,r,id,b);}
 @Get('orders/:orderId/requests/:requestId') requestView(@Param('slug') s:string,@Param('orderId') id:string,@Param('requestId') q:string,@Req() r:FastifyRequest){return this.service.requestView(s,r,id,q);}
 @Post('orders/:orderId/requests/:requestId/messages') requestReply(@Param('slug') s:string,@Param('orderId') id:string,@Param('requestId') q:string,@Req() r:FastifyRequest,@Body() b:unknown){return this.service.requestReply(s,r,id,q,b);}
 @Post('orders/:orderId/requests') request(@Param('slug') s:string,@Param('orderId') id:string,@Req() r:FastifyRequest,@Body() b:unknown){return this.service.request(s,r,id,b);}
}
@Controller('tenants/:tenantId/purchase')
@UseGuards(SessionGuard)
class MerchantController {constructor(private service:PurchaseService){}
 @Get('accounts') accounts(@Param('tenantId') t:string,@Req() r:AuthRequest){return this.service.accounts(t,r);}
 @Post('accounts/simulated') connect(@Param('tenantId') t:string,@Req() r:AuthRequest){return this.service.connect(t,r);}
 @Post('accounts/:accountId/revoke') revoke(@Param('tenantId') t:string,@Param('accountId') a:string,@Req() r:AuthRequest){return this.service.revoke(t,r,a);}
 @Get('orders') list(@Param('tenantId') t:string,@Req() r:AuthRequest){return this.service.list(t,r);}
 @Get('orders/:orderId') read(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest){return this.service.adminRead(t,r,o);}
 @Post('orders/:orderId/cancel') cancel(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest,@Body() b:unknown){return this.service.cancel(t,r,o,b);}
 @Post('orders/:orderId/address') address(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest,@Body() b:unknown){return this.service.address(t,r,o,b);}
 @Post('orders/:orderId/outbox/retry') reprocess(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest){return this.service.reprocess(t,r,o);}
 @Post('orders/:orderId/refunds') refund(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest,@Body() b:unknown){return this.service.refund(t,r,o,b);}
 @Post('orders/:orderId/reallocate') allocate(@Param('tenantId') t:string,@Param('orderId') o:string,@Req() r:AuthRequest){return this.service.reallocate(t,r,o);}
 @Post('orders/:orderId/incidents/:incidentId/note') note(@Param('tenantId') t:string,@Param('orderId') o:string,@Param('incidentId') i:string,@Req() r:AuthRequest,@Body() b:unknown){return this.service.note(t,r,o,i,b);}
 @Post('attempts/:attemptId/simulate') simulate(@Param('tenantId') t:string,@Param('attemptId') a:string,@Req() r:AuthRequest,@Body() b:unknown){return this.service.simulate(t,r,a,b);}
}
@Controller('payments')
class WebhookController {constructor(private infra:Infrastructure){}
 @Post('simulated/:accountId/webhook') webhook(@Param('accountId') a:string,@Req() r:FastifyRequest,@Body() input:unknown){const b=object(input);return receiveWebhook(this.infra.shop,id(a),text(b.event_id,100),id(b.resource_id),text(r.headers['x-request-id'],100),text(r.headers['x-signature'],200));}
}
@Module({imports:[CartModule,StoresModule,AccessModule,MfaModule],providers:[PurchaseService],controllers:[BuyerController,MerchantController,WebhookController]})
export class PurchaseModule{}