# docs/api/operacao.openapi.yaml
openapi: 3.1.0
info:
  title: E-commerce — operação do piloto (Fase 4)
  version: 1.1.0-fase4
  description: |
    Expedição, atendimento, notificações, suspensão/pausa, exportação e eliminação. Complementa compra.openapi.yaml.
    Rotas `/tenants` exigem sessão + `X-CSRF-Token`; "Dono" indica ação exclusiva do papel OWNER (403 para Funcionário).
    Erros: `{statusCode, error, requestId}`; recurso de outra loja responde 404. Paginação: no máximo 100 itens.
servers: [{url: /api}]
components:
  securitySchemes:
    session: {type: apiKey, in: cookie, name: session}
    cart: {type: apiKey, in: cookie, name: cart_<tenant>}
    orderLink: {type: apiKey, in: header, name: X-Order-Token, description: Segredo do link do e-mail (vem no fragmento `#acesso=`); hash, 90 dias, revogável.}
    supportToken: {type: apiKey, in: header, name: X-Support-Token, description: Código de acompanhamento do contato geral, exibido uma única vez.}
    exportToken: {type: apiKey, in: header, name: X-Export-Token, description: Segredo da exportação, válido por 15 minutos.}
  schemas:
    Ticket:
      type: object
      properties:
        id: {type: string, format: uuid}
        kind: {enum: [SUPPORT, WITHDRAWAL, CANCELLATION, DATA_ACCESS, DATA_ERASURE, CONTACT]}
        status: {enum: [OPEN, IN_PROGRESS, RESOLVED]}
        outcome: {enum: [ACCEPTED, DECLINED, WITHDRAWN_BY_CONSUMER, INFORMED, null]}
        created_at: {type: string, format: date-time, description: Horário original, imutável}
        due_at: {type: string, format: date-time}
        messages: {type: array, items: {type: object, properties: {author: {enum: [CONSUMER, STAFF, SYSTEM]}, body: {type: string}, created_at: {type: string}}}}
...
paths:
  /public/stores/{slug}/orders/{orderId}:
    get: {summary: Pedido (cookie do carrinho ou link do e-mail), security: [{cart: []}, {orderLink: []}], responses: {'200': {description: OK}, '404': {description: Não autorizado}}}
  /public/stores/{slug}/orders/{orderId}/requests/{requestId}:
    get: {summary: Protocolo vinculado ao pedido, security: [{cart: []}, {orderLink: []}], responses: {'200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/Ticket'}}}}}}
  /public/stores/{slug}/orders/{orderId}/requests/{requestId}/messages:
    post: {summary: Resposta do consumidor (409 se concluído), security: [{cart: []}, {orderLink: []}], requestBody: {content: {application/json: {schema: {type: object, properties: {body: {type: string, maxLength: 4000}}}}}}, responses: {'201': {description: OK}, '409': {description: Concluído}}}
...